% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{activity_features}
\alias{activity_features}
\title{List all possible activity features}
\usage{
activity_features()
}
\value{
Character vector of all possible features.
}
\description{
List all possible activity features
}
\examples{
activity_features()

## [1] "calories"  "steps"     "distance"  "floors"    "elevation" "heart"
}

