% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_functions.R
\name{activity_features_info}
\alias{activity_features_info}
\title{Get the information about the activity features.}
\usage{
activity_features_info(features = activity_features())
}
\arguments{
\item{features}{Character vector of features.}
}
\value{
Data.table with 2 columns: features and minimal_detail_level.
}
\description{
Lists the minimal `detail_level` (argument of \code{\link{get_activity}}) for each feature in `features`
}
\examples{
activity_features_info()

##     features minimal_detail_level
## 1:  calories                 1min
## 2:     steps                 1min
## 3:  distance                 1min
## 4:    floors                 1min
## 5: elevation                 1min
## 6:     heart                 1sec
}

