% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity_summary}
\alias{activity_summary}
\title{Daily Activity Summary}
\usage{
activity_summary(
  date,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{date}{The date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
Daily Activity Summary
}
\details{
See \url{https://dev.fitbit.com/build/reference/web-api/activity/} for more details.
}
