% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity_types}
\alias{activity_types}
\title{Activity Types}
\usage{
activity_types(token)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\description{
Browse Activity Types
Get a tree of all valid Fitbit public activities from the activities catalog as well as private custom activities the user created in the format requested. If the activity has levels, also get a list of activity level details.
Typically, an applications retrieve the complete list of activities once at startup to cache and show in the UI later.
}
