% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{add_alarm}
\alias{add_alarm}
\title{Add alarm}
\usage{
add_alarm(token, tracker_id, time, weekday, enabled = TRUE,
  recurring = FALSE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{tracker_id}{The ID of the tracker for which data is returned. The tracker-id value is found by get_devices().}

\item{time}{Time of day that the alarm vibrates with a UTC timezone offset, e.g. 07:15-08:00}

\item{weekday}{Comma separated list of days of the week on which the alarm vibrates, e.g. MONDAY,TUESDAY}

\item{enabled}{TRUE or FALSE. If FALSE, alarm does not vibrate until enabled is set to TRUE.}

\item{recurring}{TRUE or FALSE. If FALSE, the alarm is a single event.}
}
\description{
Adds the alarm settings to a given ID for a given device.
}
