% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{add_favorite_activity}
\alias{add_favorite_activity}
\title{Add Favorite Activity}
\usage{
add_favorite_activity(token, activity_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{activity_id}{The ID of the activity to add to user's favorites.}
}
\description{
The Add Favorite Activity endpoint adds the activity with the given ID to user's list of favorite activities.
Note: A successful request returns a 201 response code with an empty body.
}

