% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_oauth_app}
\alias{create_oauth_app}
\title{Create an OAuth app to use for authorization}
\usage{
create_oauth_app(oauth_app_name, client_id, client_secret, callback)
}
\arguments{
\item{oauth_app_name}{The name of your OAuth app. Default: \code{fitbitr}}

\item{client_id}{Your Fitbit client ID}

\item{client_secret}{Your Fitbit client secret}

\item{callback}{Your Fitbit redirect URL}
}
\description{
Create an OAuth app to use for authorization
}
