% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{delete_activity_log}
\alias{delete_activity_log}
\title{Delete Activity Log}
\usage{
delete_activity_log(token, activity_log_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{activity_log_id}{The id of the activity log entry.}
}
\description{
The Delete Activity Log endpoint deletes a user's activity log entry with the given ID. A successful request will return a 204 status code with an empty response body.
}
