% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{delete_alarm}
\alias{delete_alarm}
\title{Delete Alarm}
\usage{
delete_alarm(token, tracker_id, alarm_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{tracker_id}{The ID of the tracker whose alarms is managed. The tracker-id value is found via the Get Devices endpoint.}

\item{alarm_id}{The ID of the alarm to be updated. The alarm-id value is found in the response of the Get Alarms endpoint.}
}
\description{
Deletes the user's device alarm entry with the given ID for a given device.
}
