% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{delete_favorite_activity}
\alias{delete_favorite_activity}
\title{Delete Favorite Activity
The Delete Favorite Activity removes the activity with the given ID from a user's list of favorite activities.
Note: A successful request returns a 204 status code with an empty response body.}
\usage{
delete_favorite_activity()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{activity-id}{The ID of the activity to be removed.}
}
\description{
Delete Favorite Activity
The Delete Favorite Activity removes the activity with the given ID from a user's list of favorite activities.
Note: A successful request returns a 204 status code with an empty response body.
}

