% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{delete_sleep_log}
\alias{delete_sleep_log}
\title{Delete Sleep Log}
\usage{
delete_sleep_log(token, log_id)
}
\arguments{
\item{token}{An OAuth 2.0 token}

\item{log_id}{ID of the sleep log to be deleted.}
}
\description{
Deletes a user's sleep log entry with the given ID.
A successful request will return a 204 status code with an empty response body.
}
