% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{delete_weight_log}
\alias{delete_weight_log}
\title{Delete Weight Log
The Delete Weight Log API deletes a user's body weight log entry with the given ID.
Note: A successful request returns a 204 status code with an empty response body.}
\usage{
delete_weight_log(token, body_weight_log_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{body_weight_log_id}{The ID of the body weight log entry.}
}
\description{
Delete Weight Log
The Delete Weight Log API deletes a user's body weight log entry with the given ID.
Note: A successful request returns a 204 status code with an empty response body.
}

