% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIntraday.R
\name{extractIntraday}
\alias{extractIntraday}
\title{Takes a parsed intraday query and returns the timeseries data from within.}
\usage{
extractIntraday(parsedQuery)
}
\arguments{
\item{parsedQuery}{The query string to be sent in. See given API docs for details.}
}
\value{
dataframe with column of time, value, and type.
}
\description{
Takes a parsed intraday query and returns the timeseries data from within.
}
\examples{
extractIntraday = (myHrQuery)
}
