% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_api_functions.R
\name{fitbit_GET}
\alias{fitbit_GET}
\title{Make a GET request on \url{https://api.fitbit.com}.}
\usage{
fitbit_GET(path, token, ...)
}
\arguments{
\item{path}{Character scalar specifying the requested address from \url{https://api.fitbit.com/}.}

\item{token}{Fitbit API token generated with \code{\link{get_fitbit_token}}.}

\item{...}{Additional arguments passed to \code{\link{get_fitbit_token}}.}
}
\value{
Response object of class "response", as returned by the \code{\link[httr]{GET}} function. Use \code{\link{fitbit_parse}} to get the body of the output.
}
\description{
This is the main building block of this package. It is a wrapper for the \code{\link[httr]{GET}} function.
}
\details{
This function has one side effect: if updates the "rate_limit" attribute in the token according to the returned request header.
}
\examples{
\dontrun{
resp <- fitbit_GET("1/user/-/sleep/date/2015-12-12.json", token = fitbit_token )
ret <- fitbit_parse(resp)
}

}

