% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_api_functions.R
\name{fitbit_parse}
\alias{fitbit_parse}
\title{Parse the output of \code{\link{fitbit_GET}} using \code{\link[jsonlite]{fromJSON}}.}
\usage{
fitbit_parse(req)
}
\arguments{
\item{req}{Output object of \code{\link{fitbit_GET}}.}
}
\value{
List representing the JSON content of the GET response.
}
\description{
Parse the output of \code{\link{fitbit_GET}} using \code{\link[jsonlite]{fromJSON}}.
}

