\docType{data}
\name{fitbitData}
\alias{fitbitData}
\title{Fitbit activity (steps) data for a specific individual for 2 months.}
\format{A matrix with 288 rows and 61 columns where the rows represent specific 5 minute intervals along a day and the columns are specific days ranging from October 1st 2012 to November 30th 2012. The values are the number of steps taken by a certain individual in the specific 5 minute interval. Some days have missing data.}
\description{
  Number of steps taken by a specific individual recorded
  in 5 minute intervals for a total of 2 months using the
  Fitbit device.
}
\keyword{datasets}

