% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbitr.R
\docType{package}
\name{fitbitr}
\alias{fitbitr}
\alias{url_base}
\alias{fitbitr-package}
\title{fitbitr}
\format{An object of class \code{character} of length 1.}
\usage{
url_base
}
\description{
Provides an API Access to fitbit API via R
}
\keyword{datasets}
