% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{fitbitr_setup}
\alias{fitbitr_setup}
\title{Set Access and Refresh Token as Env Vars}
\usage{
fitbitr_setup(
  access_token = NULL,
  refresh_token = NULL,
  user_id = NULL,
  config_file = "~/.fitbitr-oauth"
)
}
\description{
Set Access and Refresh Token as Env Vars
}
