% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{fitbitr_teardown}
\alias{fitbitr_teardown}
\title{Env Teardown}
\usage{
fitbitr_teardown()
}
\value{
TRUE (invisibly)
}
\description{
Unset env variables created by `fitbitr_setup()`. Essentially, this amounts to cleaning up after yourself.
}
\examples{
\dontrun{
  fitbitr_teardown()
}
}
\seealso{
[fitbitr_setup()]
}
