% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbitTagger.R
\name{fitbitTagger}
\alias{fitbitTagger}
\title{Sets up an interface for users to tag their fitbit activities and returns a reactive dataframe that updates everytime they tag something.}
\usage{
fitbitTagger(input, output, session, data)
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}

\item{data}{dataframe returned from the function \code{\link{getPeriodProfile()}}}
}
\description{
Sets up an interface for users to tag their fitbit activities and returns a reactive dataframe that updates everytime they tag something.
}
\examples{
\dontrun{
userTags <- callModule(fitbitTagger, 'tagger', day_data = days_profile)
}
}
