% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitbitTagger.R
\name{fitbitTaggerUI}
\alias{fitbitTaggerUI}
\title{Plots a series of fitbit time series charts with a tagger.}
\usage{
fitbitTaggerUI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}

\item{...}{additional tags to \code{\link{div}} to display before plot is
displayed}
}
\value{
A large chart conataining as many days worth of data as was fed to it along with a tag interface.
}
\description{
Plots a series of fitbit time series charts with a tagger.
}
\examples{
fitbitTaggerUI('tagger')
}
