% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{floors}
\alias{floors}
\title{Floors Time Series}
\usage{
floors(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}
}
\description{
Resource path /activities/floors
}
