% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{generate_fitbitr_token}
\alias{generate_fitbitr_token}
\title{Create an object of class \code{fitbitr_token}}
\usage{
generate_fitbitr_token(
  refresh_token = Sys.getenv("FITBIT_REFRESH_TOKEN"),
  access_token = NULL,
  user_id = Sys.getenv("FITBIT_CLIENT_ID")
)
}
\arguments{
\item{refresh_token}{A refresh token. Defaults to the value of the \code{FITBIT_REFRESH_TOKEN} env var.}

\item{access_token}{An access token. Default: \code{NULL}}

\item{user_id}{A Fitbit API client ID}
}
\value{
A \code{fitbitr_token} object. Contains a single element \code{credentials} with its own two elements: \code{access_token} and \code{refresh_token}
}
\description{
This type of object can be used in place of a genuine OAuth 2.0
token for authorizing requests to the API in non-interactive sessions.
}
\details{
The class itself is quite minimal: It's just a list with two
elements: the access token (used for authorizing requests), and a refresh token.
}
