% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{perform_get}
\alias{perform_get}
\title{Perform a GET request}
\usage{
perform_get(url, ...)
}
\arguments{
\item{url}{The URL to make the request to}

\item{\dots}{Additional arguments (not currently used)}
}
\value{
The response
}
\description{
Perform a GET request
}
