% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_calories}
\alias{get_activity_calories}
\title{Activity Calories Time Series}
\usage{
get_activity_calories(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}
}
\value{
A tibble with two columns: \code{date} and \code{activity_calories}
}
\description{
Resource path /activities/activityCalories
}
\examples{
\dontrun{
start_date <- lubridate::today() - lubridate::weeks(1)
end_date <- lubridate::today()
get_activity_calories(date)
}
}
