% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_goals}
\alias{get_activity_goals}
\title{Get Activity Goals}
\usage{
get_activity_goals()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{'user-id}{The encoded ID of the user. Use "-" (dash) for current logged-in user.}

\item{period}{daily or weekly}
}
\description{
The Get Activity Goals retrieves a user's current daily or weekly activity goals using measurement units as defined in the unit system, which corresponds to the Accept-Language header provided.
Considerations
Only the current goals are returned.
Floors goal only returned for users who currently or have previously paired a Fitbit device with an altimeter.
Calories out goal represents either dynamic daily target from the Premium trainer plan or manual calorie burn goal.
The default daily goal values are:
10 floors
30 active minutes
Calories out is user specific, which accounts for BMR
}

