% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_intraday}
\alias{get_activity_intraday}
\title{Get Activity Intraday Time Series}
\usage{
get_activity_intraday(
  token,
  resource_path,
  date,
  detail_level = "15min",
  start_time = NULL,
  end_time = NULL,
  simplify = TRUE
)
}
\arguments{
\item{resource_path}{The resource path of the desired data}

\item{detail_level}{Number of data points to include. Either 1min or 15min. Optional.}

\item{start_time}{The start of the period, in the format HH:mm. Optional.}

\item{end_time}{The end of the period, in the format HH:mm. Optional.}
}
\description{
\code{get_activity_intraday()} returns intraday time series data in the specified range for a given resource.
  Access to the Intraday Time Series for personal use (accessing your own data) is available through the "Personal" App Type.
}
\details{
Available resource_path are
 \itemize{
   \item calories
   \item steps
   \item distance
   \item floors
   \item elevation
 }

 See \url{https://dev.fitbit.com/reference/web-api/activity/#get-activity-intraday-time-series} for more details.
}
