% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_intraday_time_series}
\alias{get_activity_intraday_time_series}
\title{Get Activity Intraday Time Series
Access to the Intraday Time Series for personal use (accessing your own data) is available through the "Personal" App Type.
Access to the Intraday Time Series for all other uses is currently granted on a case-by-case basis. Applications must demonstrate necessity to create a great user experience. Fitbit is very supportive of non-profit research and personal projects. Commercial applications require thorough review and are subject to additional requirements. Only select applications are granted access and Fitbit reserves the right to limit this access. To request access, email api@fitbit.com.}
\usage{
get_activity_intraday_time_series(token, resource_path, date, detail_level,
  start_time, end_time)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{resource_path}{The resource path of the desired data.}

\item{date}{The date, in the format yyyy-MM-dd or today.}

\item{detail_level}{Number of data points to include. Either 1min or 15min. Optional.}

\item{start_time}{The start of the period, in the format HH:mm. Optional.}

\item{end_time}{The end of the period, in the format HH:mm. Optional.}
}
\description{
This endpoint returns the Intraday Time Series for a given resource in the format requested. The endpoint mimics the Get Activity Time Series endpoint. If your application has the appropriate access, your calls to a time series endpoint for a specific day (by using start and end dates on the same day or a period of 1d), the response will include extended intraday values with a 1-minute detail level for that day. Unlike other time series calls that allow fetching data of other users, intraday data is available only for and to the authorized user.
If your application's http client is sensitive to the content-length of the response, you will likely prefer JSON as a response format to minimize the amount of data in the response.
If a user hasn't synced a tracker for more than seven (7) days, the tracker stores only summary values for the days older than one week. In this case, the intraday data for steps is zero and the intraday calories are calculated from BMR or EER. Intraday steps could be estimated from the total daily summary data by the equation above.
For the current day we will include only values until current timestamp (in the user's timezone.)
If a specific time interval was passed to the call via start-time/end-time parameters, data-value will include a summary only for the requested time period (versus whole day.)
For the activities/log/calories resource, each data point also includes the level field that reflects calculated activity level for that time period ( 0 - sedentary; 1 - lightly active; 2 - fairly active; 3 - very active.)
If BMR or EER have been used to estimate calories for the whole specific day (no tracker data), intraday data for calories would be flat and estimated according to the simple equation calories_1m = BMR(EER)/24/60.
Using BMR/EER Algorithms
If a user had no Fitbit tracker data for the specific day then the greater of Logged Activities + BMR (for minutes when there is no activity) and the calories calculated from the EER for that day (if EER enabled for this user's profile) are taken. In case, there was some data from the tracker for the specific day, that data where available is used and for time where data is unavailable, the BMR is used. If the total is less than 20% greater than BMR then the EER (cals < EER * 0.8) is used. EER never used to calculate calories for today.
}
\details{
Using BMR Formula
Fitbit uses the standard MD Mifflin-St Jeor equation:

9.99 * weightKg + 6.25*heightCm - 4.92*ageYears + s, where s is +5 for males and -161 for female

EER Formula (TEE total energy expenditure)

The EER Formula is based on http://www.cdc.gov/pcd/issues/2006/oct/pdf/06_0034.pdf, which in turn is based on "Food and Nutrition Board. Dietary reference intakes for energy, carbohydrate, fiber, fat, fatty acids, cholesterol, protein, and amino acids (macronutrients). Washington (DC): National Academy Press; 2005." http://www.nap.edu/openbook.php?isbn=0309085373&page=204#' MALE-based EER Formula:

TEE = 864 - 9.72 x age (years) + 1.0 x (14.2 x weight(kg) + 503 x height (meters))

FEMALE-based EER Formula:

TEE = 387 - 7.31 x age (years) + 1.0 x (10.9 x weight(kg) + 660.7 x height (meters))
}

