% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_log}
\alias{get_activity_log}
\title{Get Activity Log}
\usage{
get_activity_log(token, date = Sys.Date(), simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{A Date class object or a string in the format "yyyy-MM-dd" or "today".}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_activity_log()} retrieves a data.frame of a user's activity log entries for a given day.
}
\details{
See \url{https://dev.fitbit.com/build/reference/web-api/activity/#get-activity-logs-list} for more details.
}
