% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_logs_list}
\alias{get_activity_logs_list}
\title{Get Activity Logs List}
\usage{
get_activity_logs_list(token, before_date, after_date, sort, offset, limit)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{before_date}{The date in the format yyyy-MM-ddTHH:mm:ss. Only yyyy-MM-dd is required.
Either beforeDate or afterDate should be specified.}

\item{after_date}{The date in the format yyyy-MM-ddTHH:mm:ss.}

\item{sort}{The sort order of entries by date asc (ascending) or desc (descending).}

\item{offset}{The offset number of entries.}

\item{limit}{The max of the number of entries returned (maximum: 100)}
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-activity-logs-list} for more details.
}
