% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_summary}
\alias{get_activity_summary}
\title{Activity Summary}
\usage{
get_activity_summary(token, date)
}
\arguments{
\item{token}{A \code{fitbitr_token} object or an \code{httr::Token2.0} object a la \link[httr]{Token2.0}}

\item{date}{The date of records to be returned in "yyyy-mm-dd" or date(time) format}
}
\value{
A tibble with the \code{date} and a number of activity summary metrics for the day.
}
\description{
See \url{https://dev.fitbit.com/build/reference/web-api/activity/} for more details.
}
\examples{
\dontrun{
date <- lubridate::today()
get_activity_summary(date)
}
}
