% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_TCX}
\alias{get_activity_TCX}
\title{Get Activity TCX
Note: Since this is a beta feature, Fitbit may need to make backwards incompatible changes with less than 30 days notice.
The Get Activity TCX endpoint retrieves the details of a user's location and heart rate data during a logged exercise activity.
The Training Center XML (TCX) is a data exchange format that contains GPS, heart rate, and lap data, when it is available for the activity. The TCX MIME type is application/vnd.garmin.tcx+xml.
TCX requires GPS data and heart data points, which are tied to GPS data points. If you only want heart rate, see the heart rate intraday time series.}
\usage{
get_activity_TCX()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{user_id}{The encoded ID of the user. Use "-" (dash) for current logged-in user.}

\item{log_id}{The activity's log ID.}
}
\description{
Finding an "activity's log id"
Go to the Activity Logs page in the Fitbit Web app. Click View Details button on the Activity History pane that contains GPS data. The activity log id will be in the URL, e.g. 213560689 in https://www.fitbit.com/activities/exercise/213560689 . The GPS activities appears in the Get Activities and Get Activities List endpoints contains GPS activities.
}
