% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_time_series}
\alias{get_activity_time_series}
\title{Activity Time Series}
\usage{
get_activity_time_series(token, resource_path, date = "", period = "",
  base_date = "", end_date = "")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{resource_path}{The resource path; see options in the "Resource Path Options" section below.}

\item{date}{The end date of the period specified in the format yyyy-MM-dd or today.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, 1m, 3m, 6m, 1y, or max.}

\item{base_date}{The range start date, in the format yyyy-MM-dd or today.}

\item{end_date}{The end date of the range.}
}
\description{
The Get Activity Time Series endpoint returns time series data in the specified range for a given resource in the format requested using units in the unit system that corresponds to the Accept-Language header provided.
}
\details{
Even if you provide earlier dates in the request, the response will retrieve only data since the user's join date or the first log entry date for the requested collection.
The activities/tracker/... resource represents the daily activity values logged by the tracker device only, excluding manual activity log entries.
The activities/tracker/calories resource does not include the Estimated Energy Requirement for calorie estimation (EER) calculations for any dates even if they are turned on for the user's profile and use BMR level instead.
The activities collection is maintained as a backwards compatible resource urls (e.g. activities/log/calories).
Elevation time series (/elevation, /floors) are only available for users with compatible trackers.
}
