% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_type}
\alias{get_activity_type}
\title{Get Activity Type}
\usage{
get_activity_type(activity_id)
}
\arguments{
\item{activity_id}{The activity ID.}

\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\description{
Returns the details of a specific activity in the Fitbit activities database in the format requested.
If activity has levels, also returns a list of activity level details.
}
