% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_azm_intraday_time_series}
\alias{get_azm_intraday_time_series}
\title{Get AZM Intraday Time Series}
\usage{
get_azm_intraday_time_series(token, date,
  detail_level = "15min", start_time = NULL, end_time = NULL,
  simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The end date of the period specified. A Date class object or a string in the format yyyy-MM-dd.}

\item{detail_level}{Number of data points to include. Either 1min or 15min. Optional.}

\item{detail_level}{Number of data points to include. Either 1min or 15min. Optional.}

\item{start_time}{The start of the period, in the format HH:mm. Optional.}

\item{end_time}{The end of the period, in the format HH:mm. Optional.}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_azm_intraday_time_series()} returns intrday time series data in the specified range for active zone minutes.
}
\details{
Add details here 

 See \url{https://dev.fitbit.com/build/reference/web-api/intraday/get-azm-intraday-by-interval} for more details.
}
