% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{get_body_fat_logs}
\alias{get_body_fat_logs}
\title{Get Body Fat Logs}
\usage{
get_body_fat_logs(token, date = "", base_date = "", period = "",
  end_date = "")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The end date of the period specified in the format yyyy-MM-dd or today.}

\item{base_date}{The range start date, in the format yyyy-MM-dd or today.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, or 1m.}

\item{end_date}{The end date of the range. The period must not be longer than 31 days.}
}
\description{
The Get Body Fat Logs API retrieves a list of all user's body fat log entries for a given day in the format requested. Body fat log entries are available only to authorized user. If you need to fetch only the most recent entry, you can use the Get Body Measurements endpoint.
}
