% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_daily_activity_summary}
\alias{get_daily_activity_summary}
\title{Get Daily Activity Summary}
\usage{
get_daily_activity_summary(token, date)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The date in the format yyyy-MM-dd}
}
\description{
Retrieves a summary and list of a user's activities and activity log entries for a given day in the format requested using units in the unit system which corresponds to the Accept-Language header provided.
}
\details{
Daily summary data and daily goals for elevation (elevation, floors) only included for users with a device with an altimeter.
The steps field in activity log entires included only for activities that have steps (e.g. "Walking", "Running"); distance only included when it is relevant.
A false value for field means that activity log entry was created without explicit start time via some of our data import workflows. The field for such entry is commonly set to 00:00.
Calorie burn goal (caloriesOut) represents either dynamic daily target from the premium trainer plan or manual calorie burn goal. Goals are included to the response only for today and 21 days in the past.
}
