% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{get_devices}
\alias{get_devices}
\title{Get Devices}
\usage{
get_devices(token)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\description{
Returns a list of the Fitbit devices connected to a user's account.
}
