% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday.R
\name{get_elevation_intraday}
\alias{get_elevation_intraday}
\title{Get intraday elevationd time series}
\usage{
get_elevation_intraday(
  user_id = .fitbitr_token$credentials$user_id,
  date = lubridate::today(),
  detail_level = c("1min", "5min", "15min"),
  start_time = NULL,
  end_time = NULL
)
}
\arguments{
\item{user_id}{The fitbitr client's user id}

\item{date}{A date to get data for}

\item{detail_level}{The detail level. One of \code{"1min"}, \code{"5min"}, or \code{"15min"}}

\item{start_time}{The start time of the time window. Default: \code{NULL} gets the whole day}

\item{end_time}{The end time of the time window. Default: \code{NULL} gets the whole day}
}
\value{
A tibble with two columns: \code{time} and \code{elevation}
}
\description{
See the \href{https://dev.fitbit.com/build/reference/web-api/intraday/get-activity-intraday-by-date/}{API documentation} for
more detailed explanations of parameters and more usage information and examples.
}
\seealso{
Other intraday: 
\code{\link{get_calories_intraday}()},
\code{\link{get_distance_intraday}()},
\code{\link{get_floors_intraday}()},
\code{\link{get_steps_intraday}()}
}
\concept{intraday}
