% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_favorite_activities}
\alias{get_favorite_activities}
\title{Get Favorite Activities
Returns a list of a user's favorite activities.}
\usage{
get_favorite_activities()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{user_id}{The encoded ID of the user. Use "-" (dash) for current logged-in user.}
}
\description{
Get Favorite Activities
Returns a list of a user's favorite activities.
}

