% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart-rate.R
\name{get_heart_rate}
\alias{get_heart_rate}
\title{Heart Rate Time Series}
\usage{
get_heart_rate(
  start_date,
  end_date = NULL,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
\code{get_heart_rate()} returns time series data in the specified range
  If you specify earlier dates in the request, the response will retrieve only data since the user's join date or the first log entry date for the requested collection.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/heart-rate/#get-heart-rate-time-series} for more details.
}
