% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart-rate.R
\name{get_heart_rate_intraday_time_series}
\alias{get_heart_rate_intraday_time_series}
\title{Get Heart Rate Intraday Time Series}
\usage{
get_heart_rate_intraday_time_series(token, date, detail_level, start_time,
  end_time)
}
\arguments{
\item{date}{The date, in the format yyyy-MM-dd or today.}

\item{detail_level}{Number of data points to include. Either 1sec or 1min. Optional.}

\item{start_time}{The start of the period, in the format HH:mm. Optional.}

\item{end_time}{The end of the period, in the format HH:mm. Optional.}
}
\description{
Returns the intraday time series for a given resource in the format requested. If your application has the appropriate access, your calls to a time series endpoint for a specific day (by using start and end dates on the same day or a period of 1d), the response will include extended intraday values with a one-minute detail level for that day. Unlike other time series calls that allow fetching data of other users, intraday data is available only for and to the authorized user.
}
\details{
Access to the Intraday Time Series for personal use (accessing your own data) is available through the "Personal" App Type.
Access to the Intraday Time Series for all other uses is currently granted on a case-by-case basis.
Applications must demonstrate necessity to create a great user experience.
Fitbit is very supportive of non-profit research and personal projects.
Commercial applications require thorough review and are subject to additional requirements.
Only select applications are granted access and Fitbit reserves the right to limit this access.
To request access, email api@fitbit.com.
}

