% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_lifetime_bests}
\alias{get_lifetime_bests}
\title{Lifetime Bests}
\usage{
get_lifetime_bests(token)
}
\arguments{
\item{token}{A \code{fitbitr_token} object or an \code{httr::Token2.0} object a la \link[httr]{Token2.0}}
}
\value{
A tibble the best \code{distance}, \code{floors}, and \code{steps} (by date) tracked on any of your trackers
}
\description{
Retrieve lifetime best distance, floors, and steps
}
\examples{
\dontrun{
get_lifetime_bests()
}

}
