% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_lifetime_stats}
\alias{get_lifetime_stats}
\title{Get Lifetime Stats}
\usage{
get_lifetime_stats()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{user_id}{The encoded ID of the user. Use "-" (dash) for current logged-in user.}
}
\description{
Retrieves the user's activity statistics in the format requested using units in the unit system which corresponds to the Accept-Language header provided. Activity statistics includes Lifetime and Best achievement values from the My Achievements tile on the website dashboard. Response contains both statistics from the tracker device and total numbers including tracker data and manual activity log entries as seen on the Fitbit website dashboard.
}
\details{
Privacy Setting
The My Achievements (Friends or Anyone) privacy permission grants access to user's resource.
}

