% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_lifetime_totals}
\alias{get_lifetime_totals}
\title{Lifetime Totals}
\usage{
get_lifetime_totals()
}
\value{
A tibble of all-time totals across trackers (i.e. the total \code{distance}, \code{floors}, and \code{steps} tracked across all of your trackers)
}
\description{
Retrieve lifetime total distance, floors, and steps
}
\examples{
\dontrun{
get_lifetime_totals()
}
}
