% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_minutes_sedentary}
\alias{get_minutes_sedentary}
\title{Minutes Sedentary Time Series}
\usage{
get_minutes_sedentary(
  start_date,
  end_date,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
Resource path /activities/minutesSedentary
}
