% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{get_profile}
\alias{get_profile}
\title{Get Profile}
\usage{
get_profile(token, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_profile()} returns a user's profile.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/user/#get-profile} for more details.
}
