% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_recent_activity_types}
\alias{get_recent_activity_types}
\title{Get Recent Activity Types}
\usage{
get_recent_activity_types(token)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\description{
Retrieves a list of a user's recent activities types logged with some details of the last activity log of that type.
}
