% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{get_sleep_goal}
\alias{get_sleep_goal}
\title{Get Sleep Goal}
\usage{
get_sleep_goal(token)
}
\arguments{
\item{token}{An OAuth 2.0 token}
}
\description{
Returns a user's current sleep goal
}
\examples{
\dontrun{
#Get the current sleep goal.
get_sleep_goal(token)
}
}
