% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{get_sleep_logs}
\alias{get_sleep_logs}
\title{Get Sleep Logs
Returns a summary and list of a user's sleep log entries as well as minute by minute sleep entry data for a given day in the format requested.
The response includes summary for all sleep log entries for the given day (including naps.)
If you need to fetch data only for the user's main sleep event, you can send the request with isMainSleep=true or use a Time Series call.}
\usage{
get_sleep_logs(token, date)
}
\arguments{
\item{token}{An OAuth 2.0 token}

\item{date}{The date of records to be returned. In the format yyyy-MM-dd(character) or Date object.}
}
\value{
a list of "sleep" and "summary" data.frame. "sleep" data.frame contains time series information of sleep. "summary" data.frame contains the summary of each "sleep" information.
}
\description{
The relationship between sleep log entry properties is expressed with the following equation:
timeInBed = minutesToFallAsleep + minutesAsleep + minutesAwake + minutesAfterWakeup
}
\details{
Also, values for minuteData can be 1 ("asleep"), 2 ("awake"), or 3 ("really awake").
}

