% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{get_sleep_summary}
\alias{get_sleep_summary}
\title{Nightly Sleep Summary}
\usage{
get_sleep_summary(
  start_date,
  end_date = NULL,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\value{
a tibble of sleep summary data
}
\description{
Returns a single row tibble of summary data on the night of sleep
}
