% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{get_sleep_time_series}
\alias{get_sleep_time_series}
\title{Get Sleep Time Series}
\usage{
get_sleep_time_series(token, resource_path, date = "", period = "",
  base_date = "", end_date = "")
}
\arguments{
\item{token}{An OAuth 2.0 token}

\item{resource_path}{The resource path; see the Resource Path Options below for a list of options.}

\item{date}{The end date of the period specified in the format "yyyy-MM-dd" or "today" as a string or Date object.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, 1m, 3m, 6m, 1y, or max.}

\item{base_date}{The range start date in the format "yyyy-MM-dd" or "today" as a string or Date object.}

\item{end_date}{The end date of the range in the format "yyyy-MM-dd" or "today" as a string or Date object.}
}
\description{
Returns time series data in the specified range for a given resource in the format requested.
Note: Even if you provide earlier dates in the request, the response retrieves only data since the user's join date or the first log entry date for the requested collection.
}
\details{
Resource Path Options in more detail.
\itemize{
  \item{startTime}
  \item{timeInBed}
  \item{minutesAsleep}
  \item{awakeningsCount}
  \item{minutesAwake}
  \item{minutesToFallAsleep}
  \item{minutesAfterWakeup}
  \item{efficiency}
}
}
