% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_tracker_bests}
\alias{get_tracker_bests}
\title{Tracker Bests}
\usage{
get_tracker_bests()
}
\value{
A tibble the best \code{distance}, \code{floors}, and \code{steps} (by date) tracked on your tracker
}
\description{
Retrieve tracker best distance, floors, and steps
}
\examples{
\dontrun{
get_tracker_bests()
}

}
