% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_tracker_totals}
\alias{get_tracker_totals}
\title{Tracker Totals}
\usage{
get_tracker_totals(token)
}
\arguments{
\item{token}{A \code{fitbitr_token} object or an \code{httr::Token2.0} object a la \link[httr]{Token2.0}}
}
\value{
A tibble of all-time tracker totals (i.e. the total \code{distance}, \code{floors}, and \code{steps} tracked by your tracker)
}
\description{
Retrieve tracker total distance, floors, and steps
}
\examples{
\dontrun{
get_tracker_totals()
}

}
