% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{get_weight_logs}
\alias{get_weight_logs}
\title{Get Weight Logs
The Get Weight Logs API retrieves a list of all user's body weight log entries for a given day using units in the unit systems which corresponds to the Accept-Language header provided. Body weight log entries are available only to authorized user. Body weight log entries in response are sorted exactly the same as they are presented on the Fitbit website.}
\usage{
get_weight_logs(token, date = "", period = "", base_date = "",
  end_date = "")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The date in the format yyyy-MM-dd.}

\item{period}{The date range period. One of 1d, 7d, 30d, 1w, 1m.}

\item{base_date}{The end date when period is provided, in the format yyyy-MM-dd; range start date when a date range is provided.}

\item{end_date}{Range end date when date range is provided. Note: The period must not be longer than 31 days.}
}
\description{
Get Weight Logs
The Get Weight Logs API retrieves a list of all user's body weight log entries for a given day using units in the unit systems which corresponds to the Accept-Language header provided. Body weight log entries are available only to authorized user. Body weight log entries in response are sorted exactly the same as they are presented on the Fitbit website.
}

