% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivity.R
\name{getActivity}
\alias{getActivity}
\title{A general interface to the activity intraday api. Not typically used directly but wrapped to create more specific functions like fitrbit::getHr.}
\usage{
getActivity(token, type = "steps", date = "today", startTime = "00:00",
  endTime = "23:59", resolution = "1min")
}
\arguments{
\item{type}{Which activity api you desire. Options are calories, steps, distance, floors, and elevation.}

\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{startTime}{HH:MM 24 hour time for when you want to start getting data. Defaults to midnight.}

\item{endTime}{HH:MM 24 hour time for when you want to stop getting data. Defaults to 23:59.}

\item{resolution}{Can choose between "1min" and "15min" although I'm not sure why you'd ever want 15 minute.}

\item{config}{An oauth config object setup with your token.}
}
\value{
A dataframe with two rows. time of the day in seconds and beats per minute for that timepoint.
}
\description{
A general interface to the activity intraday api. Not typically used directly but wrapped to create more specific functions like fitrbit::getHr.
}
\examples{
# Get data on calories burned over time.
query_result <- getActivity(
  config,
  type = 'calories',
  date = date,
  startTime = startTime,
  endTime = endTime
)
}
