% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivityHeartRates.R
\name{getActivityHeartRates}
\alias{getActivityHeartRates}
\title{Takes two arrays: one of links to activity heart rate readings and another with ids for each activity and returns a long dataframe containing all the heart rate data for the supplied links.}
\usage{
getActivityHeartRates(activities_list, token)
}
\arguments{
\item{activities_list}{Dataframe returned by \code{getActiviiesList} which has the columns: \code{heartRateLink}, \code{activityName}, and \code{startTime}.}

\item{token}{Valid OAuth token for the fitbit api. Make sure you have requested the activities scope!}
}
\value{
A dataframe with columns "activeDuration": how many seconds the activity took,"activityName": fitbit name for whatever activity you did, "averageHeartRate": average heart rate during activity, "duration": how many seconds the activity lasted, "heartRateLink": link that can be used to query api for heart rate time series for this activity, "startTime": start date and time of activity; and however many rows the query returned.

A long dataframe containing all the heart rate data for the supplied links.
}
\description{
Takes two arrays: one of links to activity heart rate readings and another with ids for each activity and returns a long dataframe containing all the heart rate data for the supplied links.
}
\examples{
# Get the 100 activities leading up to april 1st 2018:
activities <- getActivitiesList(date="2018-04-01", direction = 'before', token)
# grab another batch of 100 results using supplied next link. 
more_activities <- getActivitiesList(premade_link = activites$next_link)
}
