% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDayProfile.R
\name{getDayProfile}
\alias{getDayProfile}
\title{Grabs both heartrate and steps for a given date at 1 minute intervals.}
\usage{
getDayProfile(token, date = "today")
}
\arguments{
\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{config}{An oauth config object setup with your token.}
}
\value{
A dataframe with two rows. time of the day in seconds and steps for the previous minute.
}
\description{
Grabs both heartrate and steps for a given date at 1 minute intervals.
}
\examples{
myProfile <- getDayProfile(myToken, date = "2017-07-31")
}
