% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getElevation.R
\name{getElevation}
\alias{getElevation}
\title{Grabs time series data at 1 minute intervals on elevation. It's important to note that this is a relative measure and not feet above sea-level.}
\usage{
getElevation(config, date = "today", startTime = "00:00",
  endTime = "23:59")
}
\arguments{
\item{config}{An oauth config object setup with your token.}

\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{startTime}{HH:MM 24 hour time for when you want to start getting data. Defaults to midnight.}

\item{endTime}{HH:MM 24 hour time for when you want to stop getting data. Defaults to 23:59.}
}
\value{
A dataframe with two rows. time of the day in seconds and elevation at timepoint.
}
\description{
Grabs time series data at 1 minute intervals on elevation. It's important to note that this is a relative measure and not feet above sea-level.
}
\examples{
my_elevation <- get_elevation(
  config = conf,
  date = 'today',
  startTime = "00:00",
  endTime = "23:59"
 )
}
