% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHr.R
\name{getHr}
\alias{getHr}
\title{Grabs heart-rate data from the fitbit api.}
\usage{
getHr(token, resolution = "seconds", date = "today", startTime = "00:00",
  endTime = "23:59")
}
\arguments{
\item{resolution}{Can be set to "seconds" for 1 second intervals or "minutes" for 1 minute intervals. Defaults to seconds.}

\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{startTime}{HH:MM 24 hour time for when you want to start getting data. Defaults to midnight.}

\item{endTime}{HH:MM 24 hour time for when you want to stop getting data. Defaults to 23:59.}

\item{config}{An oauth config object setup with your token.}
}
\value{
A dataframe with two rows. time of the day in seconds and beats per minute for that timepoint.
}
\description{
Grabs heart-rate data from the fitbit api.
}
\examples{
my_hr <- getHr(
  config = conf,
  resolution = 'seconds',
  date = 'today',
  startTime = "00:00",
  endTime = "23:59"
 )
}
