% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPeriodProfile.R
\name{getPeriodProfile}
\alias{getPeriodProfile}
\title{Grabs both heartrate and steps for a each day in a provided vector at 1 minute intervals.}
\usage{
getPeriodProfile(token, desired_days)
}
\arguments{
\item{desired_days}{A character vector of dates in yyyy-MM-dd format.}

\item{config}{An oauth config object setup with your token.}
}
\value{
A dataframe with two rows. time of the day in seconds and steps for the previous minute.
}
\description{
Grabs both heartrate and steps for a each day in a provided vector at 1 minute intervals.
}
\examples{
myDays <- getPeriodProfile(myToken, desired_days = c("2017-07-31", "2017-07-30", "2017-07-29"))
}
