% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSteps.R
\name{getSteps}
\alias{getSteps}
\title{Grabs time series data at 1 minute intervals on steps.}
\usage{
getSteps(token, date = "today", startTime = "00:00", endTime = "23:59")
}
\arguments{
\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{startTime}{HH:MM 24 hour time for when you want to start getting data. Defaults to midnight.}

\item{endTime}{HH:MM 24 hour time for when you want to stop getting data. Defaults to 23:59.}

\item{config}{An oauth config object setup with your token.}
}
\value{
A dataframe with two rows. time of the day in seconds and steps for the previous minute.
}
\description{
Grabs time series data at 1 minute intervals on steps.
}
\examples{
my_steps <- getSteps(
  config = conf,
  date = 'today',
  startTime = "00:00",
  endTime = "23:59"
 )
}
