% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeries.R
\name{getTimeSeries}
\alias{getTimeSeries}
\title{Queries fitbit api and returns a nice tidy dataframe of your time series for a given date and time range.}
\usage{
getTimeSeries(token, type = "heart", activityDetail = "1min",
  date = "today", startTime = "00:00", endTime = "23:59")
}
\arguments{
\item{token}{An oauth config object setup with your token.}

\item{type}{string corresponding to the type of data you want. 'heart' returns heartrate, 'steps' steps etc. For full list of possible options see the fitbit api docs (https://dev.fitbit.com/docs/activity/#get-activity-intraday-time-series).}

\item{activityDetail}{Can be set to whatever resolution you desire for data as long as the fitbit api supports it. Currently that is '1sec' and '1min' for heart rate, or '1min' '15min' for other activities.}

\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{startTime}{HH:MM 24 hour time for when you want to start getting data. Defaults to midnight.}

\item{endTime}{HH:MM 24 hour time for when you want to stop getting data. Defaults to 23:59.}
}
\value{
A dataframe with three rows. time of the day, value, and type = 'heart rate'.
}
\description{
Queries fitbit api and returns a nice tidy dataframe of your time series for a given date and time range.
}
\examples{
myHr <- getTimeSeries(
  token = conf,
  resolution = 'seconds',
  date = 'today',
  startTime = "00:00",
  endTime = "23:59"
 )
}
