% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTokenFromShiny.R
\name{getTokenFromShiny}
\alias{getTokenFromShiny}
\title{Spins up a small local shiny app that uses your fitbit api account to generate an oAuth token for you to use with the other portions of the package.}
\usage{
getTokenFromShiny(key, secret, redirect_uri,
  auth_url = "https://www.fitbit.com/oauth2/authorize",
  token_url = "https://api.fitbit.com/oauth2/token", scope = c("activity",
  "heartrate", "location", "sleep", "weight"), app_port = 5555)
}
\arguments{
\item{key}{Your apps main API Key. Refered to as 'OAuth 2.0 Client ID' on fitbits website.}

\item{secret}{Your apps secret. Longer code right below the key on the fitbit website. Called 'Client Secret'}

\item{redirect_uri}{The url that your app exists as. You should set this to wherever this app will get hosted. Most likely is http://localhost:5555 (port can be configued with argument \code{app_port})}

\item{auth_url}{Url to go to to get token. Almost always is 'https://www.fitbit.com/oauth2/authorize' so can be left at default}

\item{token_url}{Url to go to the (re)authorize token. Like \code{auth_url} this is almost always left at its default of 'https://api.fitbit.com/oauth2/token'}

\item{scope}{An array containing the data you want your app to have access to. For more details see all the options at https://dev.fitbit.com/build/reference/web-api/oauth2/#scope.}

\item{app_port}{The port you want this shiny app to use. Note that this MUST MATCH the port appended to the end of your \code{redirect_uri} that is configured on the fitbit developers site. Defaults to an arbitrary 5555 to attempt to avoid collisions with other running apps.}
}
\value{
A shiny app that will spit out text with your fitbit token once a button is clicked.
}
\description{
Spins up a small local shiny app that uses your fitbit api account to generate an oAuth token for you to use with the other portions of the package.
}
\examples{
# Spin up shiny app that will give you a token to copy
getTokenFromShiny(
  key = 'AIFHo8',
  secret = '898asdf0asdfasdff97bdaksadf',
  redirect_url = 'http://localhost:5555' 
)
}
