% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserInfo.R
\name{getUserInfo}
\alias{getUserInfo}
\title{Grabs the users info.}
\usage{
getUserInfo(token)
}
\arguments{
\item{token}{An oauth config object setup with your token.}
}
\value{
A list with lots of info about the user.
}
\description{
Grabs the users info.
}
\examples{
getUserInfo <- function(token)
}
