% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{initial_setup}
\alias{initial_setup}
\title{Persist client_id and secret in .Reviron
Simplify the setup process by persisting your Fitbit client_id and secret in the `.fitbitr-oauth` file.}
\usage{
initial_setup(
  client_id,
  client_secret,
  callback,
  path = "~/.fitbitr-oauth",
  scope = c("sleep", "activity", "heartrate", "location", "nutrition", "profile",
    "settings", "social", "weight"),
  overwrite = FALSE
)
}
\arguments{
\item{client_id}{your Fitbit client ID}

\item{client_secret}{your Fitbit client secret}

\item{callback}{your Fitbit redirect URL}

\item{path}{the path to your `.fitbitr-oauth` file}

\item{scope}{the scopes to enable}

\item{overwrite}{do you want to overwrite your existing environment variables? not recommended, as this is destructive.}
}
\value{
TRUE (invisibly) on success
}
\description{
Persist client_id and secret in .Reviron
Simplify the setup process by persisting your Fitbit client_id and secret in the `.fitbitr-oauth` file.
}
