% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{lifetime_bests}
\alias{lifetime_bests}
\title{Lifetime Bests}
\usage{
lifetime_bests(
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
Retrieve lifetime best distance, floors, and steps
}
