% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{lifetime_totals}
\alias{lifetime_totals}
\title{Lifetime Totals}
\usage{
lifetime_totals()
}
\value{
A tibble of all-time totals across trackers (i.e. the total \code{distance}, \code{floors}, \code{steps}, \code{active_score}, and \code{calories_out} tracked across all of your trackers)
}
\description{
Retrieve lifetime total distance, floors, steps calories, and active score
}
