% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{load_cached_token}
\alias{load_cached_token}
\title{Load a token from the cache}
\usage{
load_cached_token(path = ".httr-oauth")
}
\arguments{
\item{path}{the path to the file where the token is stored}
}
\value{
No return value. The token is stored in the global environment as a hidden variable named \code{.fitbitr_token}
}
\description{
Load a token from the cache
}
