% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{log_activity}
\alias{log_activity}
\title{Log Activity}
\usage{
log_activity()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{activity_id}{optional/required    This is the Activity ID of the activity, directory activity, or intensity level activity. If you pass directory activity id, Fitbit calculates and substitutes it with the corresponding intensity level activity id based on the specified distance and/or duration.}

\item{activity_name}{optional/required    Custom activity name. Either activityId or activityName must be provided.}

\item{manualCalories}{optional/required    Calories burned, specified manually. Required with activityName parameter, otherwise optional.}

\item{startTime}{required    Activity start time. Hours and minutes in the format HH:mm:ss.}

\item{durationMillis}{required    Duration in milliseconds.}

\item{date}{required    Log entry date; in the format yyyy-MM-dd.}

\item{distance}{optional/required    Distance; required for logging directory activity. In the format X.XX and in the selected distanceUnit or in the unit system that corresponds to the Accept-Language header provided.}

\item{distanceUnit}{optional    Distance measurement unit. Steps units are available only for "Walking" (activityId=90013) and "Running" (activityId=90009) directory activities and their intensity levels.}
}
\description{
The Log Activity endpoint creates log entry for an activity or user's private custom activity using units in the unit system which corresponds to the Accept-Language header provided (or using optional custom distanceUnit) and get a response in the format requested.
}
\details{
Activity Types
Any activity log entry is based on the one of activities from the catalog of activities. Though activities in catalog have a tree structure (i.e. divided into categories like "Sports and Workouts", "Running" and subcategories like "Bicycling", "Walking"), each activity leaf in the catalog can also have more granular structure. Each leaf can have a directory activity type (and have child intensity levels) or simple activity (have no levels). So, when creating a new activity log entry, use the following types of objects to passed as the activityId value:
Directory activity id: activity contains child activities which represents intensity levels (e.g. "Running")
Intensity level id: activity within directory activity (e.g. "Running/6.7mph (9 min/mile)")
Simple activity id: single activity without intensity levels (e.g. "Shoveling snow")
Note: A list of activities can be retrieved via the Browse Activities endpoint.
Custom Activities
In the process of creating new activity log entry it is possible to seamlessly create new custom activity for the user. It could be achieved using the second scenario for the set of POST parameters (providing activityName instead of activityId). In this case, either existing custom activity with the same name will be used for creating new log entry or a new record created and then logged. As for custom activities on the website, manualCalories must always provided with the activityName parameter among other required fields.

Note: The steps field is only included for activities that have steps ("Walking" or "Running") and distance is only included when it is relevant.
}

