% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{log_body_fat}
\alias{log_body_fat}
\title{Log Body Fat
The Log Body Fat API creates a log entry for body fat and returns a response in the format requested.
Note: The returned Body Fat Log IDs are unique to the user, but not globally unique.}
\usage{
log_body_fat(token, fat, date, time = "")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{fat}{Body fat; in the format X.XX}

\item{date}{Log entry date; in the format yyyy-MM-dd.}

\item{time}{optional    Time of the measurement; hours and minutes in the format HH:mm:ss, set to the last second of the day if not provided.}
}
\description{
Log Body Fat
The Log Body Fat API creates a log entry for body fat and returns a response in the format requested.
Note: The returned Body Fat Log IDs are unique to the user, but not globally unique.
}

