% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{log_sleep}
\alias{log_sleep}
\title{Log Sleep}
\usage{
log_sleep(token, startTime, duration, date)
}
\arguments{
\item{token}{An OAuth 2.0 token}

\item{startTime}{required    Start time; hours and minutes in the format HH:mm.}

\item{duration}{required    Duration in minutes.}

\item{date}{required    Log entry date in the format yyyy-MM-dd.}
}
\description{
Creates a log entry for a sleep event and returns a response in the format requested.
Keep in mind that it is NOT possible to create overlapping log entries or entries for time periods that DO NOT originate from a tracker.
Sleep log entries appear on website's sleep tracker interface according to the date on which the sleep event ends.
}
