% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{log_weight}
\alias{log_weight}
\title{Log Weight}
\usage{
log_weight(token, weight, date, time = NULL)
}
\arguments{
\item{weight}{(required) Weight - in the format X.XX.}

\item{date}{(required) Log entry date - in the format yyyy-MM-dd.}

\item{time}{(optional) Time of the measurement - hours and minutes in the format HH:mm:ss, which is set to the last second of the day if time is not provided.}
}
\description{
The Log Weight API creates log entry for a body weight using units in the unit systems that corresponds to the Accept-Language header provided and get a response in the format requested.
}
\details{
Note: The returned Weight Log IDs are unique to the user, but not globally unique.
}

