% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConfig.R
\name{makeConfig}
\alias{makeConfig}
\title{Takes a list object with three strings: API credentials of your app's name, key, and secret. Returns a config object to be passed to other api helper functions.
    If it's the first time using the function a browser will open and ask you to authorize the app. After this a token should be stored so you wont need to.}
\usage{
makeConfig(api_keys, scope = c("activity", "heartrate", "sleep", "profile"))
}
\arguments{
\item{api_keys}{An list with three values: $appname = your apps name on the api dashboard, $key = apps key (7 chars), and $secret (more than 7 chars).}

\item{scope}{Array of what parts of the api you want access to. Defaults to getting you "activity"(steps), "heartrate", and "sleep".}
}
\value{
An httr configuration object that is required anytime the fitbit api is hit.
}
\description{
Takes a list object with three strings: API credentials of your app's name, key, and secret. Returns a config object to be passed to other api helper functions.
    If it's the first time using the function a browser will open and ask you to authorize the app. After this a token should be stored so you wont need to.
}
\examples{
source("api_keys.R") #Kept somewhere super duper secret
my_config <- makeConfig(api_keys)
}
