% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQueryString.R
\name{makeQueryString}
\alias{makeQueryString}
\title{Takes a type of query, a day of the query and optionally a time range and returns the proper query string.}
\usage{
makeQueryString(type, date = "today", startTime = "00:00",
  endTime = "23:59", activityDetail = "1min")
}
\arguments{
\item{type}{What data we want. Currently supported at the default activity tags and also heart rate ('heart') and user info ('user'). See fitbit activity options for potential types https://dev.fitbit.com/docs/activity/#activity-time-series}

\item{date}{The day for which you want data. Defaults to the current day. Day format is yyyy-MM-dd.}

\item{startTime}{HH:MM 24 hour time for when you want to start getting data. Defaults to midnight.}

\item{endTime}{HH:MM 24 hour time for when you want to stop getting data. Defaults to 23:59.}
}
\value{
A dataframe with two rows. time of the day in seconds and steps for the previous minute.
}
\description{
Takes a type of query, a day of the query and optionally a time range and returns the proper query string.
}
\examples{
hrQueryString <- makeQueryString(
  type = 'heart',
  date = 'today',
  startTime = "00:00",
  endTime = "23:59"
 )
}
