% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{minutes_very_active}
\alias{minutes_very_active}
\title{Minutes Very Active Time Series}
\usage{
minutes_very_active(
  start_date,
  end_date,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
Resource path /activities/minutesVeryActive
}
