% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{oauth_token}
\alias{oauth_token}
\title{Generate an oauth token for fitbit}
\usage{
oauth_token(key = NULL, secret = NULL, callback = NULL)
}
\arguments{
\item{key}{client key generated by your own}

\item{secret}{secret key generated by your own}

\item{callback}{callback URL configured by your own}
}
\description{
Generate an oauth token for fitbit.
You have to register your own information on https://dev.fitbit.com/jp.
}
