% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{one_time_setup}
\alias{one_time_setup}
\title{Persist key and secret in .Reviron
Simplify the setup process by persisting your Fitbit key and secret in the `.Renviron` file.}
\usage{
one_time_setup(
  key,
  secret,
  callback,
  path = "~/.Renviron",
  scope = c("sleep", "activity", "heartrate", "location", "nutrition", "profile",
    "settings", "social", "weight"),
  overwrite = FALSE
)
}
\arguments{
\item{key}{your Fitbit client ID}

\item{secret}{your Fitbit client secret}

\item{callback}{your Fitbit redirect URL}

\item{path}{the path to your `.Renviron` file}

\item{scope}{the scopes to enable}

\item{overwrite}{do you want to overwrite your existing environment variables? not recommended, as this is destructive.}
}
\value{
TRUE (invisibly) on success
}
\description{
Persist key and secret in .Reviron
Simplify the setup process by persisting your Fitbit key and secret in the `.Renviron` file.
}
