% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseQuery.R
\name{parseQuery}
\alias{parseQuery}
\title{Parses a returned query from the fitbit api into a nice R object we can work with}
\usage{
parseQuery(queryResult)
}
\arguments{
\item{queryResult}{A blob of response from the fitbit api as delivered by \code{fitbitGet()}}
}
\value{
An R object with the data we desire in an easier to access setup.
}
\description{
Parses a returned query from the fitbit api into a nice R object we can work with
}
\examples{
query_string = parseQuery(fitbitGet(...))
}
