% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_api_functions.R
\name{rate_limit}
\alias{rate_limit}
\title{Get the number of requests left}
\usage{
rate_limit(token, show_message = TRUE)
}
\arguments{
\item{token}{Fitbit API token generated with \code{\link{get_fitbit_token}}.}

\item{show_message}{Boolean scalar; Should the function print the message?}
}
\value{
List containing the information about the number of requests left returned invisibly.
}
\description{
This function displays the information about how many API requests are left and
when the rate limit will be reset.
}
\examples{
\dontrun{
resp <- fitbit_GET("1/user/-/sleep/date/2015-12-12.json", token = fitbit_token )
rate_limit(fitbit_token)
## Requests left:
## 120 / 150 (reset in 22 min, at 18:00:01)
}
}

