% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{refresh_api_token.fitbitr_token}
\alias{refresh_api_token.fitbitr_token}
\title{Refresh api token}
\usage{
\method{refresh_api_token}{fitbitr_token}(
  x,
  oauth_app_name = "fitbitr",
  client_id = Sys.getenv("FITBIT_CLIENT_ID"),
  client_secret = Sys.getenv("FITBIT_CLIENT_SECRET"),
  callback = Sys.getenv("FITBIT_CALLBACK", "http://localhost:1410/"),
  use_basic_auth = TRUE
)
}
\description{
Refresh api token
}
