% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{refresh_api_token}
\alias{refresh_api_token}
\alias{refresh_fitbitr_token.default}
\title{Refresh an existing Fitbit API token}
\usage{
refresh_api_token(x, ...)

refresh_fitbitr_token.default(x, ...)
}
\description{
A token can be either of class \code{Token2.0} (from \code{httr}) or of class
\code{fitbitr_token} (from \code{fitbitr}).
}
