% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_api_functions.R
\name{refresh_fitbit_token}
\alias{refresh_fitbit_token}
\title{Refresh the expired fitbit token.}
\usage{
refresh_fitbit_token(token)
}
\arguments{
\item{token}{Fitbit API token generated with \code{\link{get_fitbit_token}}.}
}
\description{
This function refreshes the fitbit token as described in \url{https://dev.fitbit.com/docs/oauth2/#refreshing-tokens}.
The internal refresh function (\code{refresh_oauth2.0})
of the \code{\link{httr}} package somehow fails to refresh it.
}
\examples{
\dontrun{
token <- refresh_fitbit_token(token)
}
}

