% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{refresh_token}
\alias{refresh_token}
\title{Get token}
\usage{
refresh_token(
  key = Sys.getenv("FITBIT_KEY"),
  secret = Sys.getenv("FITBIT_SECRET"),
  refresh_token = Sys.getenv("FITBIT_REFRESH_TOKEN")
)
}
\description{
Get token
}
