% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGetToken.R
\name{shinyGetToken}
\alias{shinyGetToken}
\title{Helper function for retreiving token from fitbitapi in a shiny function. Used in \code{shinyLogin()}}
\usage{
shinyGetToken(auth_code, redirect_uri, key, token_url, key_secret_code)
}
\arguments{
\item{auth_code}{User's previously retreived authorization code.}

\item{redirect_uri}{your apps redirect URL (aka where you app lives on the internet or localhost)}

\item{key}{Your apps api key.}

\item{token_url}{the token retreival address for your api.}

\item{key_secret_code}{a base 64 mashup of the key and secret. Taken care of by the package}
}
\value{
A list with the response of the token post request. If no erorr will contain the access_token.
}
\description{
Helper function for retreiving token from fitbitapi in a shiny function. Used in \code{shinyLogin()}
}
\examples{
token_request <- getToken(
  auth_code = "sdfaf43qgasgadsfadsfadsfadsf",
  redirect_uri = "http://127.0.0.1:1410/",
  key = api_keys$key,
  token_url = api_keys$token_url)
}
