% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLogin.R
\name{shinyLogin}
\alias{shinyLogin}
\title{Server side component. You supply this with your api credentials.}
\usage{
shinyLogin(input, output, session, api_info, response_type = "code")
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}

\item{api_info}{A named list of the various api values you need to authenticate. "key", "id", "secret", "url", "token_url" 'https://www.fitbit.com/oauth2/authorize'}

\item{api_key}{your apps personal api key.}

\item{scope}{a vector of what you're requesting access to in the API. See the given api docs for examples.}
}
\description{
Server side component. You supply this with your api credentials.
}
\examples{
authButton <- callModule(shinyauth, "fitbit_login", api_info = api_keys)
}
