% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLogin.R
\name{shinyLoginUI}
\alias{shinyLoginUI}
\title{Allows the easy retreival of a token for using apis in your shiny apps.}
\usage{
shinyLoginUI(id)
}
\arguments{
\item{id}{the id you will use to keep track of this component in your app}
}
\value{
A button that a used presses to login to your given authentication service.
}
\description{
Allows the easy retreival of a token for using apis in your shiny apps.
}
\examples{
shinydrawrUI('myrecorder')
}
