% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleep_analyzer}
\alias{sleep_analyzer}
\title{Analyze sleep patterns}
\usage{
sleep_analyzer(start_date = "2017-01-01", end_date = "2018-06-14",
  planned_start = 22, planned_end = 8, limit = NULL, contrast = NULL,
  smooth_span = 0.25, bar_spacing = 0.1)
}
\arguments{
\item{start_date}{An Retrieve fitbit data from (e.g. "2017-01-01")}

\item{end_date}{Retrieve fitbit data to (e.g. "2018-06-14")}

\item{planned_start}{At what time do you plan to sleep usually (e.g. 22)?}

\item{limit}{Can be NULL, weekdays (e.g. 'Monday'), or Months (e.g. 'July'). With 'limit' you can limit your results but it can also be the comparison for contrast}

\item{contrast}{Can be NULL, weekdays (e.g. 'Monday'), or Months (e.g. 'July').Can be used to compare sleep durations with values specified in 'limit'}

\item{smooth_span}{Plot smoothing parameter}

\item{bar_spacing}{Plot style paramter}

\item{end_date}{When do you (plan to) get up (e.g. 8)?}
}
\description{
Retrieves and analyzes sleep patterns
}
