% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleep_summary}
\alias{sleep_summary}
\title{Nightly Sleep Summary}
\usage{
sleep_summary(start_date, end_date = start_date)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}
}
\value{
a tibble of sleep summary data
}
\description{
Returns a tibble of summary by night
}
