% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{tracker_totals}
\alias{tracker_totals}
\title{Tracker Totals}
\usage{
tracker_totals(
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
Retrieve tracker total distance, floors, steps calories, and active score
}
