% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{update_activity_goals}
\alias{update_activity_goals}
\title{Update Activity Goals}
\usage{
update_activity_goals()
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\description{
The Update Activity Goals endpoint creates or updates a user's daily activity goals and returns a response using units in the unit system which corresponds to the Accept-Language header provided.
user-id	The encoded ID of the user. Use "-" (dash) for current logged-in user.
period	daily or weekly
caloriesOut	optional	Goal value; integer.
activeMinutes	optional	Goal value; integer.
floors	optional	Goal value; integer.
distance	optional	Goal value; in the format X.XX or integer.
steps	optional	Goal value; integer.
}

