% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{update_alarm}
\alias{update_alarm}
\title{Update Alarm}
\usage{
update_alarm(token, tracker_id, alarm_id, time, weekday, enabled = TRUE,
  recurring = FALSE, snooze_length = 3, snooze_count = 9,
  label = "", vibe = "DEFAULT")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{tracker_id}{The ID of the tracker whose alarms is managed. The tracker-id value is found via the Get Devices endpoint.}

\item{alarm_id}{The ID of the alarm to be updated. The alarm-id value is found in the response of the Get Alarms endpoint.}

\item{time}{Time of day that the alarm vibrates with a UTC timezone offset, e.g. 07:15-08:00}

\item{weekday}{Comma separated list of days of the week on which the alarm vibrates, e.g. MONDAY,TUESDAY}

\item{enabled}{TRUE or FALSE. If FALSE, alarm does not vibrate until enabled is set to TRUE.}

\item{recurring}{TRUE or FALSE. If FALSE, the alarm is a single event.}

\item{snooze_length}{Minutes between alarms; integer value.}

\item{snooze_count}{Maximum snooze count; integer value.}

\item{label}{Label for the alarm; string value.}

\item{vibe}{Vibe pattern; only one value for now - DEFAULT.}
}
\description{
Updates the alarm entry with a given ID for a given device. It also gets a response in the format requested.
}
