% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{update_body_fat_goal}
\alias{update_body_fat_goal}
\title{Update Body Fat Goal}
\usage{
update_body_fat_goal(token, goal)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{fat}{(required) Target body fat percentage; in the format X.XX.}
}
\description{
The Update Body Fat Goal API creates or updates user's fat percentage goal.
}

