% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{update_body_weight_goal}
\alias{update_body_weight_goal}
\title{Update Weight Goal
The Update Weight Goal API creates or updates user's fat or weight goal using units in the unit systems that corresponds to the Accept-Language header provided in the format requested.}
\usage{
update_body_weight_goal(token, startDate, startWeight, weight = "")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{startDate}{Weight goal start date; in the format yyyy-MM-dd.}

\item{startWeight}{Weight goal start weight; in the format X.XX, in the unit systems that corresponds to the Accept-Language header provided.}

\item{weight}{required/optional    Weight goal target weight; in the format X.XX, in the unit systems that corresponds to the Accept-Language header provided; required if user doesn't have an existing weight goal.}
}
\description{
Update Weight Goal
The Update Weight Goal API creates or updates user's fat or weight goal using units in the unit systems that corresponds to the Accept-Language header provided in the format requested.
}

