% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{update_sleep_goal}
\alias{update_sleep_goal}
\title{Update Sleep Goal}
\usage{
update_sleep_goal(token, min_duration)
}
\arguments{
\item{token}{An OAuth 2.0 token}

\item{min_duration}{The target sleep duration is in minutes.}
}
\description{
The Update Sleep Goal endpoint creates or updates a user's sleep goal and get a response in the in the format requested
}
\examples{
\dontrun{
#Set a new sleep goal(377)
update_sleep_goal(token, 377)
}
}
