% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity_summary}
\alias{activity_summary}
\title{Activity Summary}
\usage{
activity_summary(date)
}
\arguments{
\item{date}{The date of records to be returned in "yyyy-mm-dd" or date(time) format}
}
\value{
A tibble with the \code{date} and a number of activity summary metrics for the day.
}
\description{
See \url{https://dev.fitbit.com/build/reference/web-api/activity/} for more details.
}
