% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{add_favorite_activity}
\alias{add_favorite_activity}
\title{Add Favorite Activity}
\usage{
add_favorite_activity(token, activity_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{activity_id}{The ID of the activity to add to user's favorites.}
}
\description{
\code{add_favorite_activity()} adds the activity with the given ID to user's list of favorite activities.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#add-favorite-activity} for more details.
}
