% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_oauth_app}
\alias{create_oauth_app}
\title{Create an OAuth app to use for authorization}
\usage{
create_oauth_app(oauth_app_name, client_id, client_secret, callback)
}
\description{
Create an OAuth app to use for authorization
}
