% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{delete_body_fat_log}
\alias{delete_body_fat_log}
\title{Delete Body Fat Log}
\usage{
delete_body_fat_log(token, log_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{log_id}{The ID of the body fat log entry.}
}
\description{
The Delete Body Fat Log API deletes a user's body fat log entry with the given ID.
}
\details{
Note: A successful request returns a 204 status code with an empty response body.
}
