% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{delete_favorite_activity}
\alias{delete_favorite_activity}
\title{Delete Favorite Activity}
\usage{
delete_favorite_activity(token, activity_id)
}
\arguments{
\item{activity_id}{The ID of the activity to be removed.}
}
\description{
\code{delete_favorite_activity()} removes the activity with the given ID (activity_id) from a user's list of favorite activities.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#delete-favorite-activity} for more details.
}
