% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{delete_weight_log}
\alias{delete_weight_log}
\title{Delete Weight Log}
\usage{
delete_weight_log(token, log_id)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{log_id}{The ID of the weight log entry.}
}
\description{
\code{delete_weight_log()} deletes a user's body weight log entry with the given ID.
}
