% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{fitbitr_setup}
\alias{fitbitr_setup}
\title{Set Access and Refresh Token as Env Vars}
\usage{
fitbitr_setup(
  access_token = NULL,
  refresh_token = NULL,
  user_id = NULL,
  config_file = "~/.fitbitr-oauth"
)
}
\arguments{
\item{access_token}{the Fitbit access token (default: NULL)}

\item{refresh_token}{the Fitbit refresh token (default: NULL)}

\item{user_id}{the Fitbit user_id (default: NULL)}

\item{config_file}{the path to your config file (default: `~/.fitbitr-oauth`)}
}
\value{
TRUE (invisibly) on success
}
\description{
Set Access and Refresh Token as Env Vars
}
