% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{fitbitr_setup}
\alias{fitbitr_setup}
\title{Set necessary env vars}
\usage{
fitbitr_setup(
  config_file = "~/.fitbitr-oauth",
  access_token = NULL,
  refresh_token = NULL,
  user_id = NULL,
  client_id = NULL,
  client_secret = NULL
)
}
\arguments{
\item{config_file}{the path to your config file (default: `~/.fitbitr-oauth`)}

\item{access_token}{the Fitbit access token (default: NULL)}

\item{refresh_token}{the Fitbit refresh token (default: NULL)}

\item{user_id}{the Fitbit user_id (default: NULL)}

\item{client_id}{the Fitbit Client Id (default: NULL)}

\item{client_secret}{the Fitbit Client Secret (default: NULL)}
}
\value{
TRUE (invisibly) on success
}
\description{
Make calling the Fitbit API easier by setting env vars at setup time.
This function sets the config that's necessary for making API requests,
including the access token, refresh token, user id, client id, and client secret.
Then, functions making API calls will use `Sys.getenv()` to access the config.
}
