% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get}
\alias{get}
\title{Perform a GET request}
\usage{
get(token, url, ...)
}
\arguments{
\item{token}{A \code{fitbitr_token} object or an \code{httr::Token2.0} object a la \link[httr]{Token2.0}}

\item{url}{The URL to make the request to}

\item{\dots}{Additional arguments (not currently used)}
}
\value{
The response
}
\description{
Perform a GET request
}
