% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_goals}
\alias{get_activity_goals}
\title{Get Activity Goals}
\usage{
get_activity_goals(token, period, simplify = TRUE)
}
\arguments{
\item{period}{"daily" or "weekly"}
}
\description{
\code{get_activity_goals()} retrieves a user's current daily or weekly activity goals
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-activity-goals} for more details.
}
