% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_intraday_time_series}
\alias{get_activity_intraday_time_series}
\title{Get Activity Intraday Time Series}
\usage{
get_activity_intraday_time_series(token, resource_path, date,
  detail_level = "15min", start_time = NULL, end_time = NULL,
  simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{resource_path}{The resource path of the desired data}

\item{date}{A Date class object or a string in the format "yyyy-MM-dd" or "today".}

\item{detail_level}{Number of data points to include. Either 1min or 15min. Optional.}

\item{start_time}{The start of the period, in the format HH:mm. Optional.}

\item{end_time}{The end of the period, in the format HH:mm. Optional.}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_activity_intraday_time_series()} returns intraday time series data in the specified range for a given resource.
  Access to the Intraday Time Series for personal use (accessing your own data) is available through the "Personal" App Type.
}
\details{
Available resource_path are
 \itemize{
   \item calories
   \item steps
   \item distance
   \item floors
   \item elevation
 }

 See \url{https://dev.fitbit.com/reference/web-api/activity/#get-activity-intraday-time-series} for more details.
}
