% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_log}
\alias{get_activity_log}
\title{Get Activity Log}
\usage{
get_activity_log(token, date = Sys.Date(), simplify = TRUE)
}
\description{
\code{get_activity_log()} retrieves a data.frame of a user's activity log entries for a given day.
}
\details{
See \url{https://dev.fitbit.com/build/reference/web-api/activity/#get-activity-logs-list} for more details.
}
