% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_logs_list}
\alias{get_activity_logs_list}
\title{Get Activity Logs List}
\usage{
get_activity_logs_list(before_date, after_date, sort, offset, limit)
}
\arguments{
\item{sort}{The sort order of entries by date asc (ascending) or desc (descending).}

\item{offset}{The offset number of entries.}

\item{limit}{The max of the number of entries returned (maximum: 100)}

\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{beforeDate}{The date in the format yyyy-MM-ddTHH:mm:ss. Only yyyy-MM-dd is required. Either beforeDate or afterDate should be specified.}

\item{afterDate}{The date in the format yyyy-MM-ddTHH:mm:ss.}
}
\description{
Retrieves a list of a user's activity log entries before or after a given day with offset and limit using units in the unit system which corresponds to the Accept-Language header provided.
Note: This is a beta feature that will soon have a backwards incompatible change. Please see this announcement.
}

