% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_summary}
\alias{get_activity_summary}
\title{Get Daily Activity Summary}
\usage{
get_activity_summary(token, date, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{A Date class object or a string in the format "yyyy-MM-dd" or "today".}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_activity_summary()} retrieves a summary and list of a user's activities for a given day.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-daily-activity-summary} for more details.
}
