% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_time_series}
\alias{get_activity_time_series}
\title{Activity Time Series}
\usage{
get_activity_time_series(
  start_date,
  end_date,
  resource_path,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{resource_path}{The resource path. See \url{https://dev.fitbit.com/build/reference/web-api/activity/} for options}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\description{
Activity Time Series
}
\details{
See \url{https://dev.fitbit.com/build/reference/web-api/activity/} for more details.
}
