% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_time_series}
\alias{get_activity_time_series}
\title{Get Activity Time Series}
\usage{
get_activity_time_series(token, resource_path, date = "", period = "",
  base_date = "", end_date = "", simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{resource_path}{The resource path. see details below.}

\item{date}{The end date of the period specified. A Date class object or a string in the format yyyy-MM-dd.}

\item{period}{The range for which data will be returned. Options are "1d", "7d", "30d", "1w", "1m", "3m", "6m", "1y", or "max".}

\item{base_date}{The range start date. A Date class object or a string in the format yyyy-MM-dd or today.}

\item{end_date}{The end date of the range. A Date class object or a string in the format yyyy-MM-dd.}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_activity_time_series()} returns time series data in the specified range for a given resource.
}
\details{
Available resource_path are
 \itemize{
   \item calories
   \item caloriesBMR
   \item steps
   \item distance
   \item floors
   \item elevation
   \item minutesSedentary
   \item minutesLightlyActive
   \item minutesFairlyActive
   \item minutesVeryActive
   \item activityCalories
 }

 See \url{https://dev.fitbit.com/reference/web-api/activity/#get-activity-time-series} for more details.
}
