% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_type}
\alias{get_activity_type}
\title{Get Activity Type}
\usage{
get_activity_type(token, activity_id, simplify = TRUE)
}
\arguments{
\item{activity_id}{The activity ID.}
}
\description{
\code{get_activity_type()} returns the details of a specific activity in the Fitbit activities database.
}
