% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_type}
\alias{get_activity_type}
\title{Get Activity Type}
\usage{
get_activity_type(token, activity_id, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{activity_id}{The activity ID.}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_activity_type()} returns the details of a specific activity in the Fitbit activities database.
}
