% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_activity_types}
\alias{get_activity_types}
\title{Get Activity Types}
\usage{
get_activity_types(token, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
Get a tree of all valid Fitbit public activities from the activities catalog as well as private custom activities the user created
}
