% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{get_body_fat_logs}
\alias{get_body_fat_logs}
\title{Get Body Fat Logs}
\usage{
get_body_fat_logs(
  token,
  date = "",
  base_date = "",
  period = "",
  end_date = "",
  simplify = TRUE
)
}
\arguments{
\item{date}{The end date of the period specified in the format yyyy-MM-dd or today.}

\item{base_date}{The range start date, in the format yyyy-MM-dd or today.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, or 1m.}

\item{end_date}{The end date of the range. The period must not be longer than 31 days.}
}
\description{
\code{get_body_fat_logs()} retrieves a list of all user's body fat log entries for a given day.
}
