% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{get_body_fat_logs}
\alias{get_body_fat_logs}
\title{Get Body Fat Logs
The Get Body Fat Logs API retrieves a list of all user's body fat log entries for a given day in the format requested. Body fat log entries are available only to authorized user. If you need to fetch only the most recent entry, you can use the Get Body Measurements endpoint.}
\usage{
get_body_fat_logs(token, date = "", period = "", base_date = "",
  end_date = "")
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The date in the format yyyy-MM-dd.}

\item{period}{The date range period. One 1d, 7d, 1w, 1m.}

\item{base_date}{The end date when period is provided; range start date when a date range is provided. In the format yyyy-MM-dd or today.}

\item{end_date}{Range end date when date range is provided. Note: The range should not be longer than 31 days}
}
\description{
Get Body Fat Logs
The Get Body Fat Logs API retrieves a list of all user's body fat log entries for a given day in the format requested. Body fat log entries are available only to authorized user. If you need to fetch only the most recent entry, you can use the Get Body Measurements endpoint.
}

