% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R, R/body_weight.R
\name{get_body_goals}
\alias{get_body_goals}
\alias{get_body_goals}
\title{Get Body Goals}
\usage{
get_body_goals(token, goal_type)

get_body_goals(token, goal_type)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{goal_type}{weight or fat}

\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{goal_type}{weight or fat}
}
\description{
The Get Body Goals API retrieves a user's current body fat percentage or weight goal using units in the unit systems that corresponds to the Accept-Language header provided in the format requested.

\code{get_body_goals()} retrieves a user's current body fat percentage or weight goal.
}
