% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{get_body_goals}
\alias{get_body_goals}
\title{Get Body Goals}
\usage{
get_body_goals(token, goal_type)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{goal_type}{weight or fat}
}
\description{
\code{get_body_goals()} retrieves a user's current body fat percentage or weight goal.
}
