% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{get_body_time_series}
\alias{get_body_time_series}
\title{Get Body Time Series}
\usage{
get_body_time_series(token, resource_path, base_date = NULL,
  end_date = NULL, date = NULL, period = NULL)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{resource_path}{The resource path. Options are "bmi", "fat", or "weight".}

\item{base_date}{The range start date, in the format yyyy-MM-dd or today.}

\item{end_date}{The end date of the range.}

\item{date}{The end date of the period specified in the format yyyy-MM-dd or today.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, 1m, 3m, 6m, 1y, or max.}
}
\description{
\code{get_body_time_series()} returns time series data in the specified range for a given resource.
}
