% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_favorite_activities}
\alias{get_favorite_activities}
\title{Get Favorite Activities}
\usage{
get_favorite_activities(token)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\description{
Returns a list of a user's favorite activities.
}

