% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_favorite_activities}
\alias{get_favorite_activities}
\title{Get Favorite Activities}
\usage{
get_favorite_activities(token, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_favorite_activities()} returns a list of a user's favorite activities.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-favorite-activities} for more details.
}
