% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_frequent_activities}
\alias{get_frequent_activities}
\title{Get Frequent Activities}
\usage{
get_frequent_activities(token)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-frequent-activities}  for more details.
}
