% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_frequent_activities}
\alias{get_frequent_activities}
\title{Get Frequent Activities}
\usage{
get_frequent_activities(token, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_frequent_activities()} retrieves a list of a user's frequent activities.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-frequent-activities}  for more details.
}
