% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heart-rate.R
\name{get_heart_rate_intraday_time_series}
\alias{get_heart_rate_intraday_time_series}
\title{Get Heart Rate Intraday Time Series}
\usage{
get_heart_rate_intraday_time_series(token, date = "",
  detail_level = "1min", start_time = NULL, end_time = NULL,
  simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The end date of the period specified in the format "yyyy-MM-dd" or "today" as a string or Date object.}

\item{detail_level}{Number of data points to include. Either "1sec" or "1min".}

\item{start_time}{The start of the period, in the format "HH:mm".}

\item{end_time}{The end of the period, in the format "HH:mm"}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_heart_rate_intraday_time_series()} returns the intraday time series.
  If your application has the appropriate access, your calls to a time series endpoint for a specific day (by using start and end dates on the same day or a period of 1d),
  the response will include extended intraday values with a one-minute detail level for that day.
  Access to the Intraday Time Series for personal use (accessing your own data) is available through the "Personal" App Type.
}
\details{
See \url{'https://dev.fitbit.com/reference/web-api/heart-rate/#get-heart-rate-intraday-time-series} for more details.
}
