% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_lifetime_stats}
\alias{get_lifetime_stats}
\title{Get Lifetime Stats}
\usage{
get_lifetime_stats(token, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_lifetime_stats()} retrieves the user's activity statistics.
  Activity statistics includes Lifetime and Best achievement values from the My Achievements tile on the website dashboard.
  Response contains both statistics from the tracker device and total numbers including tracker data and manual activity log entries as seen on the Fitbit website dashboard.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-lifetime-stats} for more details.
}
