% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_recent_activity_types}
\alias{get_recent_activity_types}
\title{Get Recent Activity Types}
\usage{
get_recent_activity_types()
}
\description{
Retrieves a list of a user's recent activities types logged with some details of the last activity log of that type using units in the unit system which corresponds to the Accept-Language header provided. The record retrieved can be used to log the activity via the Log Activity endpoint with the same or adjusted values for distance and duration.
}

