% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{get_recent_activity_types}
\alias{get_recent_activity_types}
\title{Get Recent Activity Types}
\usage{
get_recent_activity_types(token, simplify = TRUE)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{simplify}{logical; should the result be simplified from a list to a data.frame if possible}
}
\description{
\code{get_recent_activity_types()} retrieves a list of a user's recent activities types logged with some details of the last activity log of that type.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#get-recent-activity-types} for more details.
}
