% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body.R
\name{get_weight_logs}
\alias{get_weight_logs}
\title{Get Weight Logs}
\usage{
get_weight_logs(token, date = NULL, base_date = NULL, period = NULL,
  end_date = NULL)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The end date of the period specified in the format yyyy-MM-dd or today.}

\item{base_date}{The range start date, in the format yyyy-MM-dd or today.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, or 1m.}

\item{end_date}{The end date of the range. The period must not be longer than 31 days.}
}
\description{
The Get Weight Logs endpoint returns time series data in the specified range for a given resource in the format requested using units in the unit system that corresponds to the Accept-Language header provided.
Even if you provide earlier dates in the request, the response will retrieve only data since the user's join date or the first log entry date for the requested collection.
}
