% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{get_weight_logs}
\alias{get_weight_logs}
\title{Get Weight Logs}
\usage{
get_weight_logs(token, date = NULL, base_date = NULL, period = NULL,
  end_date = NULL)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{date}{The end date of the period specified in the format yyyy-MM-dd or today.}

\item{base_date}{The range start date, in the format yyyy-MM-dd or today.}

\item{period}{The range for which data will be returned. Options are 1d, 7d, 30d, 1w, or 1m.}

\item{end_date}{The end date of the range. The period must not be longer than 31 days.}
}
\description{
\code{get_weight_logs()} returns time series data in the specified range for a given resource.
}
