% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUserInfo.R
\name{getUserInfo}
\alias{getUserInfo}
\title{Grabs the users info and returns a giant list of things to see. For the full list of things see (https://dev.fitbit.com/docs/user/). Each key is available as the corresponding key of the R return object.}
\usage{
getUserInfo(token)
}
\arguments{
\item{token}{An oauth config object setup with your token.}
}
\value{
A list with lots of info about the user.
}
\description{
Grabs the users info and returns a giant list of things to see. For the full list of things see (https://dev.fitbit.com/docs/user/). Each key is available as the corresponding key of the R return object.
}
\examples{
getUserInfo <- function(token)
}
