% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{lifetime_bests}
\alias{lifetime_bests}
\title{Lifetime Bests}
\usage{
lifetime_bests()
}
\value{
A tibble the best \code{distance}, \code{floors}, and \code{steps} (by date) tracked on any of your trackers
}
\description{
Retrieve lifetime best distance, floors, and steps
}
