% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{log_body_fat}
\alias{log_body_fat}
\title{Log Body Fat}
\usage{
log_body_fat(token, fat, date, time = NULL)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{fat}{Body fat; in the format X.XX, in the unit system that corresponds to the Accept-Language header provided.}

\item{date}{Log entry date; in the format yyyy-MM-dd.}

\item{time}{Time of the measurement; hours and minutes in the format HH:mm:ss, set to the last second of the day if not provided.}
}
\description{
\code{log_body_fat()} creates a log entry for body fat.
}
