% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_weight.R
\name{log_weight}
\alias{log_weight}
\title{Log Weight}
\usage{
log_weight(token, weight, date, time = NULL)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{weight}{Weight - in the format X.XX.}

\item{date}{Log entry date - in the format yyyy-MM-dd.}

\item{time}{Time of the measurement - hours and minutes in the format HH:mm:ss, which is set to the last second of the day if time is not provided.}
}
\description{
\code{log_weight()} creates log entry for a body weight and get a response in the format requested.
}
