% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{oauth_token}
\alias{oauth_token}
\title{Generate an oauth token for fitbit}
\usage{
oauth_token(key = NULL, secret = NULL, callback = NULL)
}
\arguments{
\item{key}{client key generated by your own}

\item{secret}{secret key generated by your own}

\item{callback}{callback URL configured by your own}
}
\description{
OAuth token will be generated by "Client" as OAuth 2.0 Application Type.
If you want to use "Personal" type, plesase set key and secret variable. In this case, you have to register your own application on https://dev.fitbit.com/jp.
}
