% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{refresh_token}
\alias{refresh_token}
\title{Get token}
\usage{
refresh_token(
  refresh_token = Sys.getenv("FITBIT_REFRESH_TOKEN"),
  client_id = Sys.getenv("FITBIT_CLIENT_ID"),
  client_secret = Sys.getenv("FITBIT_CLIENT_SECRET"),
  expires_in = 86400
)
}
\arguments{
\item{refresh_token}{the Fitbit refresh token}

\item{client_id}{the Fitbit Client ID}

\item{client_secret}{the Fitbit Client Secret}

\item{expires_in}{How long do you want your token to remain valid? (default: 86400 sec = 1 day)}
}
\value{
TRUE (invisibly) on success
}
\description{
Get token
}
