% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLogin.R
\name{shinyLogin}
\alias{shinyLogin}
\title{Server side component. You supply this with your api credentials.}
\usage{
shinyLogin(input, output, session, api_info, response_type = "code")
}
\arguments{
\item{input}{you can ignore this as it is taken care of by shiny}

\item{output}{you can ignore this as it is taken care of by shiny}

\item{session}{you can ignore this as it is taken care of by shiny}

\item{api_info}{A named list of the various api values you need to authenticate. 
'key': string of app ID, 
'secret': long string of secret, 
'auth_url': Where the app needs to go for authenticating: usually is 'https://www.fitbit.com/oauth2/authorize',
"redirect_uri": URL that the api will send back to after authentication: aka your app's url,
"token_url": URL for renewing a stale token. Usually 'https://api.fitbit.com/oauth2/token',
'scope': an array of desired data from api. See here https://dev.fitbit.com/build/reference/web-api/oauth2/#scope for options.}

\item{scope}{a vector of what you're requesting access to in the API. See the given api docs for examples.}
}
\description{
Server side component. You supply this with your api credentials.
}
\examples{
authButton <- callModule(shinyLogin, "fitbit_login", api_info = api_keys)
}
