% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleep_stage_granular}
\alias{sleep_stage_granular}
\title{Granular Sleep Stage Data}
\usage{
sleep_stage_granular(
  start_date,
  end_date = start_date,
  token = Sys.getenv("FITBIT_ACCESS_TOKEN"),
  user_id = Sys.getenv("FITBIT_USER_ID")
)
}
\arguments{
\item{start_date}{The start date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{end_date}{The end date of records to be returned in "yyyy-mm-dd" or date(time) format}

\item{token}{Fitbit access token}

\item{user_id}{Fitbit user id}
}
\value{
a tibble of sleep summary data
}
\description{
Returns a tibble of nightly sleep stage data.
Very granular. Returns blocks of time spent in each phase.
}
