% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{tracker_totals}
\alias{tracker_totals}
\title{Tracker Totals}
\usage{
tracker_totals()
}
\value{
A tibble of all-time tracker totals (i.e. the total \code{distance}, \code{floors}, and \code{steps} tracked by your tracker)
}
\description{
Retrieve tracker total distance, floors, and steps
}
\examples{
\dontrun{
tracker_totals()
}
}
