% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{update_activity_goals}
\alias{update_activity_goals}
\title{Update Activity Goals}
\usage{
update_activity_goals(token, period, calories_out = NULL,
  active_minutes = NULL, floors = NULL, distance = NULL,
  steps = NULL)
}
\arguments{
\item{token}{An OAuth 2.0 token generated by oauth_token()}

\item{period}{daily or weekly}

\item{calories_out}{optional    Goal value; integer.}

\item{active_minutes}{optional    Goal value; integer.}

\item{floors}{optional    Goal value; integer.}

\item{distance}{optional    Goal value; in the format X.XX or integer.}

\item{steps}{optional    Goal value; integer.}
}
\description{
\code{update_activity_goals()} creates or updates a user's daily activity goals.
}
\details{
See \url{https://dev.fitbit.com/reference/web-api/activity/#update-activity-goals} for more details.
}
